fix(proxy): split rust control-plane auth and logging endpoints
